Catch up to main (#8)
* Change charger and battery_level columns from integer to double (#5)

Tesla Fleet Telemetry sends float values for charger_power (e.g. 1.4 kW),
charger_voltage (e.g. 113.78V), charger_current, and battery_level
(e.g. 32.05%). SQLite masks this with dynamic typing, but the integer
declaration would truncate values on MySQL/PostgreSQL. The battery_level
integer model cast was also silently losing precision.

Fixes #4

* Fix --after/--before date parsing in process-states command (#6)

* Parse --after/--before through Carbon::parse()->utc() in process-states

ISO 8601 strings (with 'T' separator) were compared directly against
SQLite timestamps (space separator). Since space < 'T' in ASCII, the
string comparison silently failed — charges whose started_at fell
within the window were not matched for deletion.

Same fix pattern as 3886a9a (API date filtering).

* Convert parsed dates to strings to match processVehicle(?string) signature

+++ app/Models/VehicleState.php
@@ -48,7 +48,7 @@ class VehicleState extends Model
             'locked' => 'boolean',
             'sentry_mode' => 'boolean',
             'climate_on' => 'boolean',
-            'battery_level' => 'integer',
+            'battery_level' => 'float',
             'rated_range' => 'float',
         ];
     }
